+++ sojo-app/app/controllers/index.js
@@ -7,18 +7,6 @@ export default Controller.extend({
       return this.get ('gatekeeper.isSignedIn');
   }),
   actions: {
-      createAccount () {
-      let {email, username, password} = this.getProperties (['email', 'username', 'password']);
-      let account = this.get ('store').createRecord ('account', {username, password, email});
-      let adapterOptions = {signIn: true};
-
-      account.save ({adapterOptions}).then (account => {
-        // You can transition to a protected application route
-      }).catch (reason => {
-        // Display error message to user
-      });
-    },
-
     signOut() {
       this.get('gatekeeper').signOut().then(() => {
         this.replaceRoute('login');
